Approach 3: Route groups + CSS visibility toggle
