+++ calculator-app/src/app/[countryId]/(with-sidebar)/households/page.tsx
@@ -1,20 +1,12 @@
 "use client";
 
 import { useEffect } from "react";
-import StandardLayout from "@/components/StandardLayout";
 import PopulationsPage from "@/pages/Populations.page";
 import { perfContentVisible } from "@/utils/perfHarness";
-import { CalculatorProviders } from "../providers";
 
 export default function HouseholdsRoute() {
   // [PERF HARNESS]
   useEffect(() => { perfContentVisible('HouseholdsRoute'); }, []);
 
-  return (
-    <CalculatorProviders>
-      <StandardLayout>
-        <PopulationsPage />
-      </StandardLayout>
-    </CalculatorProviders>
-  );
+  return <PopulationsPage />;
 }
