+++ calculator-app/src/app/[countryId]/(with-sidebar)/policies/page.tsx
@@ -1,20 +1,12 @@
 "use client";
 
 import { useEffect } from "react";
-import StandardLayout from "@/components/StandardLayout";
 import PoliciesPage from "@/pages/Policies.page";
 import { perfContentVisible } from "@/utils/perfHarness";
-import { CalculatorProviders } from "../providers";
 
 export default function PoliciesRoute() {
   // [PERF HARNESS]
   useEffect(() => { perfContentVisible('PoliciesRoute'); }, []);
 
-  return (
-    <CalculatorProviders>
-      <StandardLayout>
-        <PoliciesPage />
-      </StandardLayout>
-    </CalculatorProviders>
-  );
+  return <PoliciesPage />;
 }
