+++ calculator-app/src/app/[countryId]/(with-sidebar)/reports/create/page.tsx
@@ -1,20 +1,12 @@
 "use client";
 
 import { useEffect } from "react";
-import StandardLayout from "@/components/StandardLayout";
 import ReportBuilderPage from "@/pages/reportBuilder/ReportBuilderPage";
 import { perfContentVisible } from "@/utils/perfHarness";
-import { CalculatorProviders } from "../../providers";
 
 export default function ReportBuilderRoute() {
   // [PERF HARNESS]
   useEffect(() => { perfContentVisible('ReportBuilderRoute'); }, []);
 
-  return (
-    <CalculatorProviders>
-      <StandardLayout>
-        <ReportBuilderPage />
-      </StandardLayout>
-    </CalculatorProviders>
-  );
+  return <ReportBuilderPage />;
 }
