+++ calculator-app/src/app/[countryId]/(with-sidebar)/reports/page.tsx
@@ -1,20 +1,12 @@
 "use client";
 
 import { useEffect } from "react";
-import StandardLayout from "@/components/StandardLayout";
 import ReportsPage from "@/pages/Reports.page";
 import { perfContentVisible } from "@/utils/perfHarness";
-import { CalculatorProviders } from "../providers";
 
 export default function ReportsRoute() {
   // [PERF HARNESS]
   useEffect(() => { perfContentVisible('ReportsRoute'); }, []);
 
-  return (
-    <CalculatorProviders>
-      <StandardLayout>
-        <ReportsPage />
-      </StandardLayout>
-    </CalculatorProviders>
-  );
+  return <ReportsPage />;
 }
