+++ calculator-app/src/app/[countryId]/(with-sidebar)/simulations/page.tsx
@@ -1,20 +1,12 @@
 "use client";
 
 import { useEffect } from "react";
-import StandardLayout from "@/components/StandardLayout";
 import SimulationsPage from "@/pages/Simulations.page";
 import { perfContentVisible } from "@/utils/perfHarness";
-import { CalculatorProviders } from "../providers";
 
 export default function SimulationsRoute() {
   // [PERF HARNESS]
   useEffect(() => { perfContentVisible('SimulationsRoute'); }, []);
 
-  return (
-    <CalculatorProviders>
-      <StandardLayout>
-        <SimulationsPage />
-      </StandardLayout>
-    </CalculatorProviders>
-  );
+  return <SimulationsPage />;
 }
