+++ calculator-app/src/app/[countryId]/layout.tsx
@@ -3,15 +3,17 @@
 import { use, useEffect, useMemo } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import { CountryProvider } from "@/contexts/CountryContext";
+import { LayoutVisibilityProvider } from "@/contexts/LayoutVisibilityContext";
 import { LocationProvider } from "@/contexts/LocationContext";
 import { NavigationProvider } from "@/contexts/NavigationContext";
 import { countryIds, type CountryId } from "@/libs/countries";
 import { perfNavChange } from "@/utils/perfHarness";
+import { CalculatorProviders } from "./providers";
 
 /**
  * Layout for extracted Next.js pages under /:countryId/*.
- * Provides CountryContext, NavigationContext, and LocationContext
- * so shared components work identically in both router contexts.
+ * Provides CountryContext, NavigationContext, LocationContext,
+ * and CalculatorProviders. StandardLayout is in the route group layouts.
  */
 export default function CountryLayout({
   children,
@@ -62,7 +64,11 @@ export default function CountryLayout({
     <CountryProvider value={countryId as CountryId}>
       <NavigationProvider value={navValue}>
         <LocationProvider value={locationValue}>
-          {children}
+          <CalculatorProviders>
+            <LayoutVisibilityProvider>
+              {children}
+            </LayoutVisibilityProvider>
+          </CalculatorProviders>
         </LocationProvider>
       </NavigationProvider>
     </CountryProvider>
