--- calculator-app/src/app/[countryId]/loading.tsx
@@ -1,17 +1,0 @@
-export default function CountryLoading() {
-  return (
-    <div style={{ display: "flex", alignItems: "center", justifyContent: "center", height: "100vh" }}>
-      <div
-        style={{
-          width: 32,
-          height: 32,
-          border: "3px solid #e5e7eb",
-          borderTopColor: "#0d9488",
-          borderRadius: "50%",
-          animation: "spin 0.6s linear infinite",
-        }}
-      />
-      <style>{`@keyframes spin { to { transform: rotate(360deg) } }`}</style>
-    </div>
-  );
-}
